+++ model_navigator/pipelines/builders/find_device_max_batch_size.py
@@ -12,40 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Find Max Batch size pipelines builders."""
-from typing import Dict, List, Union
+from typing import Dict, List
 
-from model_navigator.api.config import DeviceKind, Format
+from model_navigator.api.config import Format
 from model_navigator.commands.base import ExecutionUnit
 from model_navigator.commands.find_max_batch_size import FindMaxBatchSize, FindMaxBatchSizeConfig
 from model_navigator.configuration.common_config import CommonConfig
-from model_navigator.configuration.model.model_config import (
-    ModelConfig,
-    TensorFlowTensorRTConfig,
-    TensorRTConfig,
-    TorchTensorRTConfig,
-)
+from model_navigator.configuration.model.model_config import ModelConfig
 from model_navigator.core.logger import LOGGER
 from model_navigator.frameworks import Framework
 from model_navigator.pipelines.pipeline import Pipeline
 from model_navigator.runners.onnx import OnnxrtCUDARunner
 from model_navigator.runners.tensorflow import TensorFlowSavedModelCUDARunner
 from model_navigator.runners.torch import TorchScriptCUDARunner
-
-
-def do_run_max_batch_size_search(
-    config: CommonConfig,
-    model_cfg: Union[TensorRTConfig, TensorFlowTensorRTConfig, TorchTensorRTConfig],
-) -> bool:
-    """Should max batch size search be run for the model.
-
-    Args:
-        config: Common optimize configuration.
-        model_cfg: Model configuration.
-
-    Returns:
-        bool: True if run max batch size.
-    """
-    return bool(model_cfg.trt_profiles) is False and config.batch_dim is not None
+from model_navigator.utils.config_helpers import do_find_device_max_batch_size
 
 
 def find_device_max_batch_size_builder(
@@ -62,22 +42,8 @@ def find_device_max_batch_size_builder(
     """
     pipeline_name = "Find Device Max Batch Size"
     execution_units: List[ExecutionUnit] = []
-    model_formats = models_config.keys()
-    adaptive_formats = {Format.TORCH_TRT, Format.TENSORRT, Format.TF_TRT}
-    matching_formats = adaptive_formats.intersection(set(model_formats))
 
-    if len(matching_formats) == 0 or config.target_device != DeviceKind.CUDA:
-        LOGGER.debug("No matching formats found")
-        return Pipeline(name=pipeline_name, execution_units=execution_units)
-
-    run_search = False
-    for fmt in adaptive_formats:
-        for model_cfg in models_config.get(fmt, []):
-            if do_run_max_batch_size_search(config, model_cfg):
-                run_search = True
-
-    if not run_search:
-        LOGGER.debug("Run search disabled.")
+    if not do_find_device_max_batch_size(config=config, models_config=models_config):
         return Pipeline(name=pipeline_name, execution_units=execution_units)
 
     configurations = []