+++ tests/functional/L1_Optimize_wo_MaxBS_and_MaxPerf_Search_jax_GPT2_HuggingFace/test.sh
@@ -20,11 +20,7 @@ export XLA_PYTHON_CLIENT_PREALLOCATE=false
 THIS_SCRIPT_PATH="$(realpath --relative-to="$(pwd)" "$0")"
 TEST_MODULE="$(dirname "${THIS_SCRIPT_PATH}"|sed 's/\//./g').test"
 
-#JAX_VERSION="jax[cuda11_pip]"
-JAX_VERSION="jax[cuda12_pip]<0.4.14"
-
-pip install --upgrade "${JAX_VERSION}" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
-pip install transformers flax
+pip install --upgrade transformers flax "jax[cuda12_pip]<0.4.16" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
 
 python -m"${TEST_MODULE}" \
     --status $(pwd)/status.yaml \