+++ CHANGELOG.md
@@ -24,6 +24,7 @@ limitations under the License.
 - fix: Use wide range of shapes for TensorRT conversion
 - fix: Sorting of samples loaded from workspace
 - change: in Inplace, store one sample by default per module and store shape info for all samples
+- change: include additional export formats when max batch size search is enabled
 
 - Known issues and limitations:
   - nav.Module moves original torch.nn.Module to the CPU, in case of weight sharing that might result in unexpected behaviour